Removed assert statement as sometimes the process doesn't exist. This should be ok we just can't try and kill a process that doesn't exist I think.

+++ src/tracker_heartbeat.py
@@ -29,10 +29,13 @@ def terminate_process_and_children(p):
     ps_command = subprocess.Popen("ps -o pid --ppid %d --noheaders" % p.pid, shell=True, stdout=subprocess.PIPE)
     ps_output = ps_command.stdout.read()
     retcode = ps_command.wait()
-    assert retcode == 0, "ps command returned %d" % retcode
-    for pid_str in ps_output.split("\n")[:-1]:
-        os.kill(int(pid_str), signal.SIGINT)
-    p.terminate()
+    # assert retcode == 0, "ps command returned %d" % retcode
+    if retcode == 0:
+        rospy.loginfo("Process doesn't exist, ps command returned %d", retcode)
+    else:
+        for pid_str in ps_output.split("\n")[:-1]:
+            os.kill(int(pid_str), signal.SIGINT)
+        p.terminate()
 
 class Heartbeat_Monitor:
     def __init__(self):
